# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgentFlow AI Agent Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, draft]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create a new agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentBody"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent by ID
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Agent not found
    put:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentBody"
      responses:
        "200":
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Agent deleted
...
  /workflows/{workflowId}/run:
    post:
      operationId: runWorkflow
      tags: [executions]
      summary: Execute a workflow
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunWorkflowBody"
      responses:
        "201":
          description: Execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
  /executions:
    get:
      operationId: listExecutions
      tags: [executions]
      summary: List all executions
      parameters:
        - name: workflowId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of executions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionList"
  /executions/{executionId}:
    get:
      operationId: getExecution
      tags: [executions]
      summary: Get execution details
      parameters:
        - name: executionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Execution details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
  /executions/{executionId}/cancel:
    post:
      operationId: cancelExecution
      tags: [executions]
      summary: Cancel a running execution
      parameters:
        - name: executionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Execution cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
  /executions/{executionId}/logs:
    get:
      operationId: getExecutionLogs
      tags: [executions]
      summary: Get execution step logs
      parameters:
        - name: executionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Execution logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExecutionLog"
...
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Get analytics overview
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: week
      responses:
        "200":
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
  /analytics/execution-stats:
    get:
      operationId: getExecutionStats
      tags: [analytics]
      summary: Get execution statistics over time
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: week
      responses:
        "200":
          description: Execution statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExecutionStat"
components:
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        role:
          type: string
        goal:
          type: string
        backstory:
          type: string
        model:
          type: string
        provider:
          type: string
          enum: [openai, anthropic, google, local]
        temperature:
          type: number
        maxTokens:
          type: integer
        tools:
          type: array
          items:
            type: string
        memoryEnabled:
          type: boolean
        status:
          type: string
          enum: [active, inactive, draft]
        systemPrompt:
          type: string
        icon:
          type: string
        color:
          type: string
        executionCount:
          type: integer
        avgResponseTime:
          type: number
        successRate:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - model
        - provider
        - status
        - createdAt
        - updatedAt
    CreateAgentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        role:
          type: string
        goal:
          type: string
        backstory:
          type: string
        model:
          type: string
        provider:
          type: string
          enum: [openai, anthropic, google, local]
        temperature:
          type: number
        maxTokens:
          type: integer
        tools:
          type: array
          items:
            type: string
        memoryEnabled:
          type: boolean
        systemPrompt:
          type: string
        icon:
          type: string
        color:
          type: string
      required:
        - name
        - role
        - model
        - provider
    UpdateAgentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        role:
          type: string
        goal:
          type: string
        backstory:
          type: string
        model:
          type: string
        provider:
          type: string
          enum: [openai, anthropic, google, local]
        temperature:
          type: number
        maxTokens:
          type: integer
        tools:
          type: array
          items:
            type: string
        memoryEnabled:
          type: boolean
        status:
          type: string
          enum: [active, inactive, draft]
        systemPrompt:
          type: string
        icon:
          type: string
        color:
          type: string
...
    RunWorkflowBody:
      type: object
      properties:
        inputData:
          type: object
        config:
          type: object
    Execution:
      type: object
      properties:
        id:
          type: integer
        workflowId:
          type: integer
        workflowName:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        inputData:
          type: object
        outputData:
          type: object
        error:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        duration:
          type: number
        tokensUsed:
          type: integer
        cost:
          type: number
        currentStep:
          type: string
        totalSteps:
          type: integer
        completedSteps:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - workflowId
        - status
        - createdAt
    ExecutionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Execution"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset
    ExecutionLog:
      type: object
      properties:
        id:
          type: integer
        executionId:
          type: integer
        nodeId:
          type: string
        nodeName:
          type: string
        nodeType:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, skipped]
        input:
          type: object
        output:
          type: object
        error:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        duration:
          type: number
        tokensUsed:
          type: integer
      required:
        - id
        - executionId
        - nodeId
        - status
...
    AnalyticsOverview:
      type: object
      properties:
        totalAgents:
          type: integer
        activeAgents:
          type: integer
        totalWorkflows:
          type: integer
        activeWorkflows:
          type: integer
        totalExecutions:
          type: integer
        successfulExecutions:
          type: integer
        failedExecutions:
          type: integer
        avgExecutionTime:
          type: number
        totalTokensUsed:
          type: integer
        totalCost:
          type: number
        executionTrend:
          type: number
        successRateTrend:
          type: number
      required:
        - totalAgents
        - activeAgents
        - totalWorkflows
        - activeWorkflows
        - totalExecutions
        - successfulExecutions
        - failedExecutions
    ExecutionStat:
      type: object
      properties:
        date:
          type: string
        total:
          type: integer
        successful:
          type: integer
        failed:
          type: integer
        avgDuration:
          type: number
        tokensUsed:
          type: integer
      required:
        - date
        - total
        - successful
        - failed