# lib/api-zod/src/generated/types/createAgentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFlow AI Agent Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAgentBodyProvider } from "./createAgentBodyProvider";

export interface CreateAgentBody {
  name: string;
  description?: string;
  role: string;
  goal?: string;
  backstory?: string;
  model: string;
  provider: CreateAgentBodyProvider;
  temperature?: number;
  maxTokens?: number;
  tools?: string[];
  memoryEnabled?: boolean;
  systemPrompt?: string;
  icon?: string;
  color?: string;
}