# lib/api-zod/src/generated/types/executionLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFlow AI Agent Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionLogInput } from "./executionLogInput";
import type { ExecutionLogOutput } from "./executionLogOutput";
import type { ExecutionLogStatus } from "./executionLogStatus";

export interface ExecutionLog {
  id: number;
  executionId: number;
  nodeId: string;
  nodeName?: string;
  nodeType?: string;
  status: ExecutionLogStatus;
  input?: ExecutionLogInput;
  output?: ExecutionLogOutput;
  error?: string;
  startedAt?: Date;
  completedAt?: Date;
  duration?: number;
  tokensUsed?: number;
}