# lib/api-zod/src/generated/types/executionStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFlow AI Agent Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ExecutionStat {
  date: string;
  total: number;
  successful: number;
  failed: number;
  avgDuration?: number;
  tokensUsed?: number;
}